⚙️ Consolidate glossary foundation boundaries
