+++ bridge/app/lib/src/foundation/restricted_file_writer.dart
